refactor(create_store_controller): change the methods in branch controller to store controller
