+++ src/controllers/BranchController.java
@@ -8,28 +8,6 @@ import models.Product;
 import models.Store;
 
 public class BranchController {
-    public String[] getBranchesAsHTMLTemplate() {
-        ArrayList<Branch> branches = Store.getBranches();
-        Integer numberOfBranches = Store.getBranches().size();
-
-        String branchesAsHTMLTemplate[] = new String[numberOfBranches];
-        for (Integer index = 0; index < numberOfBranches; index++) {
-            Branch branch = branches.get(index);
-            String HTMLTemplate = String.format("""
-                    <html>
-                        <body>
-                            UUID: %s
-                            <br>
-                            Endereço: %s
-                        </body>
-                    </html>
-                    """, branch.getId(), branch.getAddress().toString());
-            branchesAsHTMLTemplate[index] = HTMLTemplate;
-        }
-
-        return branchesAsHTMLTemplate;
-    }
-
     public String[] searchProductsByWordAsHTMLTemplate(String branchUUID, String word) {
         Branch branch = getBranchByUUID(branchUUID);
         ArrayList<Product> searchedProducts = branch.searchProductsByWord(word);
@@ -61,7 +39,6 @@ public class BranchController {
     public Branch getBranchByUUID(String branchUUID) {
         ArrayList<Branch> branches = Store.getBranches();
         Branch branch = branches.stream().filter(b -> b.getId().equals(branchUUID)).findFirst().orElse(null);
-
         return branch;
     }
 
@@ -87,17 +64,5 @@ public class BranchController {
         if (branch != null) {
             branch.logout();
         }
-    }
-
-    public boolean authenticateStore(String password){
-        if (password.equals(Store.getpassword())){
-            return true;
-        } else return false;
-    }
-
-    public void removeBranch(String branchUUID) {
-
-        Branch branch = getBranchByUUID(branchUUID);
-        Store.removeBranch(branch);
     }
 }
